requirements.txt
matplotlib==3.6.0
Pillow==9.3.0
